Created new branch for imagenet development.
The main file to investigate is titled trial2.py. Within, a VGG16 net is created with imagenet weights which are frozen, then a dense layer for the A and B bands are fed on as the classifier. This is once again a very rough outline and does not currently operate at perfection.