More changes to imagenet. Began creating loop to gather all input images
The framework for a loop to gather all input images has been added. It basically needs to be repeated for a validation set. The model fit line needs to be changed to accept the new lists of input and a newly created validation set. Then its off to the races.