Added generator like features to imagenet branch.
There is a lot of cleaning up to do in this working piece of code, I wanted to push it so the group could see the entire thought process. I've commented out the possibly useless pieces of code to be removed once we can verify this is working. Currently, the code tests for 100 epochs and takes an eternity. Could maybe switch network models to one which trains faster, or find another way. I created a generator get_images which yields what the model expects for an input. Perhaps I am missing something but it doesn't work very well. It fails upon finding an image which does not have 3 bands of color, or at least that's how I interpreted the error message.